Fix: 피드백 반영 - SharedPreferences 리팩토링
